+++ src/cli/display/aligned_print.rs
@@ -1,30 +1,32 @@
 // SPDX-License-Identifier: GPL-3.0-only
 use std::fmt::Display;
 
-pub const VERTICAL_LINE_PREFIX: &str = " \u{2502} ";
-
 pub struct PairAligner {
     pairs: Vec<(String, String)>,
 }
 
 impl PairAligner {
+    pub const VERTICAL_LINE_PREFIX: &str = " \u{2502} ";
+
     pub fn new() -> Self {
         Self { pairs: Vec::new() }
     }
 
+    /// Creates a `PairAligner` from a `Vec` of pairs which have `impl Display` for their members.
     #[expect(unused)]
     pub fn from(pairs: Vec<(impl Display, impl Display)>) -> Self {
-        let pairs = pairs.into_iter().map(|p| (p.0.to_string(), p.1.to_string())).collect();
+        let pairs = pairs.into_iter().map(|(key, value)| (key.to_string(), value.to_string())).collect();
         Self { pairs }
     }
 
+    /// Adds a pair which has `impl Display` for it's members to the stored pairs.
     pub fn add(&mut self, key: impl Display, value: impl Display) {
         self.pairs.push((key.to_string(), value.to_string()));
     }
 
-    /// Prints key value pairs in an aligned format.
+    /// Prints key value pairs in an aligned format. With the given prefix in front.
     pub fn display(&self, prefix: &str) {
-        let Some(longest_key) = self.pairs.iter().max_by_key(|s| s.0.len()) else {
+        let Some(longest_key) = self.pairs.iter().max_by_key(|(key, _)| key.len()) else {
             return;
         };